+++ bahmnicore-omod/pom.xml
@@ -60,11 +60,6 @@
             <artifactId>communication-api</artifactId>
         </dependency>
         <dependency>
-            <groupId>org.bahmni.module</groupId>
-            <artifactId>fhir-ts-services-int-api</artifactId>
-            <version>${terminologyServices.version}</version>
-        </dependency>
-        <dependency>
             <groupId>org.openmrs.module</groupId>
             <artifactId>fhir2-api</artifactId>
             <version>${openmrs.fhir2.version}</version>
