+++ bahmnicore-omod/src/main/resources/config.xml
@@ -34,7 +34,6 @@
         <require_module>org.openmrs.module.auditlog</require_module>
         <require_module>org.bahmni.module.bahmnicommons</require_module>
         <require_module>org.bahmni.module.communication</require_module>
-        <require_module>org.bahmni.module.fhirterminologyservices</require_module>
     </require_modules>
     <!-- Extensions -->
