+++ pom.xml
@@ -70,7 +70,6 @@
         <providermanagementVersion>2.13.0</providermanagementVersion>
         <velocityVersion>1.6.2</velocityVersion>
         <bahmniCommons.version>1.0.0</bahmniCommons.version>
-        <terminologyServices.version>1.0.0-SNAPSHOT</terminologyServices.version>
         <openmrs.fhir2.version>1.8.0</openmrs.fhir2.version>
 
         <!--test dependencies-->
@@ -279,13 +278,6 @@
                 <artifactId>bahmni-commons-api</artifactId>
                 <type>jar</type>
                 <version>${bahmniCommons.version}</version>
-                <scope>provided</scope>
-            </dependency>
-            <dependency>
-                <groupId>org.bahmni.module</groupId>
-                <artifactId>fhir-ts-services-int-api</artifactId>
-                <type>jar</type>
-                <version>${terminologyServices.version}</version>
                 <scope>provided</scope>
             </dependency>
             <dependency>
